fix: apply LLM cleanup and dictionary replacements to streaming transcription
The streaming (WebSocket) path was sending raw transcription text directly
to the client, skipping LLM post-processing, dictionary replacements, and
per-app format rules that the REST path applied. This also meant history
entries from streaming were missing cleaned_text, LLM metadata, and cost.

Extract post-processing logic from transcribe.ts into a shared post-process
module and call it from both the REST and streaming paths. Add a 'context'
WebSocket message type so the client can forward app context for format rule
matching.
