+++ apps/server/src/routes/transcribe.ts
@@ -1,87 +1,13 @@
-import { generateText, experimental_transcribe as transcribe } from "ai";
+import { experimental_transcribe as transcribe } from "ai";
 import { Hono } from "hono";
 import { getDb } from "../lib/db.js";
+import { postProcess } from "../lib/post-process.js";
 import {
-  createChatModel,
   createTranscriptionModel,
   getDefaultModels,
 } from "../lib/providers.js";
-import { getModelCost } from "../routes/models.js";
 
 const transcribeRoute = new Hono();
-
-// ---------------------------------------------------------------------------
-// Context detection — uses format_rules from DB
-// ---------------------------------------------------------------------------
-
-/** Build a context string from the raw x-app-context header for matching */
-function buildMatchContext(rawContext: string | null): string {
-  if (!rawContext) return "";
-
-  try {
-    const ctx = JSON.parse(rawContext) as {
-      app?: string;
-      url?: string;
-      title?: string;
-      windowTitle?: string;
-    };
-
-    // Build a combined string for pattern matching
-    const parts: string[] = [];
-    if (ctx.url) parts.push(ctx.url);
-    if (ctx.title) parts.push(ctx.title);
-    if (ctx.windowTitle) parts.push(ctx.windowTitle);
-    if (ctx.app) parts.push(ctx.app);
-    return parts.join(" ");
-  } catch {
-    return rawContext;
-  }
-}
-
-/** Look up formatting instructions from the format_rules table */
-function getContextHint(
-  rawContext: string | null,
-  db: ReturnType<typeof getDb>,
-): string {
-  if (!rawContext) return "";
-
-  const matchStr = buildMatchContext(rawContext);
-  if (!matchStr) return "";
-
-  try {
-    // User rules (is_default=0) first, then defaults (is_default=1)
-    const rows = db
-      .prepare(
-        "SELECT app_pattern, instructions FROM format_rules ORDER BY is_default ASC, id DESC",
-      )
-      .all() as { app_pattern: string; instructions: string }[];
-
-    for (const row of rows) {
-      const patterns = row.app_pattern.split("|").map((p) => p.trim());
-      for (const pattern of patterns) {
-        if (pattern && matchStr.toLowerCase().includes(pattern.toLowerCase())) {
-          return row.instructions;
-        }
-      }
-    }
-  } catch {
-    // format_rules table may not exist yet
-  }
-
-  // Fallback: extract app name for a generic hint
-  try {
-    const ctx = JSON.parse(rawContext) as { app?: string };
-    if (ctx.app) return `The user is dictating in ${ctx.app}.`;
-  } catch {
-    // not JSON
-  }
-
-  return "";
-}
-
-// ---------------------------------------------------------------------------
-// Route
-// ---------------------------------------------------------------------------
 
 transcribeRoute.post("/", async (c) => {
   const start = Date.now();
@@ -121,7 +47,6 @@ transcribeRoute.post("/", async (c) => {
 
   // Step 1: Transcribe
   const db = getDb();
-  const contextHint = getContextHint(appContext, db);
   let rawText: string;
 
   const langSetting = db
@@ -159,103 +84,9 @@ transcribeRoute.post("/", async (c) => {
     });
   }
 
-  // Step 2: LLM post-processing (optional)
-  let cleanedText = rawText;
-  let inputTokens = 0;
-  let outputTokens = 0;
-
-  const llmSetting = db
-    .prepare("SELECT value FROM settings WHERE key = 'llm_cleanup'")
-    .get() as { value: string } | undefined;
-  const llmEnabled = llmSetting?.value === "true";
-
-  if (llmEnabled && defaults.llm) {
-    const systemPrompt = `You are an intelligent voice-to-text post-processor that transforms raw dictated speech into clean, polished writing.
-${contextHint ? `\nContext: ${contextHint}\n` : ""}
-Your job:
-- Remove filler words (um, uh, like, you know, basically, so, I mean, etc.)
-- Remove false starts, repeated words, and self-corrections (keep only the final intended version)
-- Fix grammar, spelling, punctuation, and capitalization
-- Convert spoken numbers, dates, and abbreviations to their written forms where appropriate
-- Structure run-on sentences into clear, well-punctuated prose
-- Preserve the speaker's original meaning, intent, tone, and personality exactly
-- Keep technical terms, names, and domain-specific vocabulary intact
-- Do NOT add information that wasn't spoken
-- Do NOT change the meaning or rewrite beyond what's needed for clarity
-- Do NOT add greetings, sign-offs, or any framing text
-
-Output ONLY the cleaned text. No explanations, no quotes, no prefixes.`;
-
-    try {
-      const chatModel = createChatModel(
-        defaults.llm.provider,
-        defaults.llm.model_id,
-      );
-      const result = await generateText({
-        model: chatModel,
-        system: systemPrompt,
-        prompt: rawText,
-      });
-      cleanedText = result.text;
-      inputTokens = result.usage?.inputTokens ?? 0;
-      outputTokens = result.usage?.outputTokens ?? 0;
-    } catch (err) {
-      console.error("LLM cleanup failed:", err);
-    }
-  }
-
-  // Step 3: Dictionary replacements
-  try {
-    const dictRows = db
-      .prepare(
-        "SELECT id, key, value FROM dictionary ORDER BY length(key) DESC",
-      )
-      .all() as { id: number; key: string; value: string }[];
-
-    if (dictRows.length > 0) {
-      const matchedIds: number[] = [];
-      for (const { id, key, value } of dictRows) {
-        const escaped = key.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-        const regex = new RegExp(`\\b${escaped}\\b`, "gi");
-        if (regex.test(cleanedText)) {
-          matchedIds.push(id);
-          cleanedText = cleanedText.replace(
-            new RegExp(`\\b${escaped}\\b`, "gi"),
-            value,
-          );
-        }
-      }
-      if (matchedIds.length > 0) {
-        const updateStmt = db.prepare(
-          "UPDATE dictionary SET usage_count = usage_count + 1 WHERE id = ?",
-        );
-        for (const id of matchedIds) {
-          updateStmt.run(id);
-        }
-      }
-    }
-  } catch {
-    // Dictionary table may not exist yet
-  }
-
+  // Step 2: LLM post-processing + dictionary replacements
+  const pp = await postProcess(rawText, appContext);
   const durationMs = Date.now() - start;
-
-  // Calculate cost from models.dev pricing
-  let costUsd = 0;
-  if (inputTokens > 0 || outputTokens > 0) {
-    try {
-      const llmModelId =
-        llmEnabled && defaults.llm ? defaults.llm.model_id : null;
-      if (llmModelId) {
-        const pricing = await getModelCost(llmModelId);
-        if (pricing) {
-          costUsd = inputTokens * pricing.input + outputTokens * pricing.output;
-        }
-      }
-    } catch {
-      // ignore pricing errors
-    }
-  }
 
   // Save to history
   try {
@@ -265,15 +96,15 @@ Output ONLY the cleaned text. No explanations, no quotes, no prefixes.`;
        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
     ).run(
       rawText,
-      cleanedText !== rawText ? cleanedText : null,
+      pp.cleaned !== rawText ? pp.cleaned : null,
       defaults.voice.provider,
       defaults.voice.model_id,
-      llmEnabled && defaults.llm ? defaults.llm.provider : null,
-      llmEnabled && defaults.llm ? defaults.llm.model_id : null,
+      pp.llmProvider,
+      pp.llmModel,
       durationMs,
-      inputTokens,
-      outputTokens,
-      costUsd,
+      pp.inputTokens,
+      pp.outputTokens,
+      pp.costUsd,
     );
   } catch (err) {
     console.error("Failed to save history:", err);
@@ -281,7 +112,7 @@ Output ONLY the cleaned text. No explanations, no quotes, no prefixes.`;
 
   return c.json({
     raw: rawText,
-    cleaned: cleanedText,
+    cleaned: pp.cleaned,
     model: defaults.voice.model_id,
     durationMs,
   });
